requirements.txt
Flask~=2.0.3
requests~=2.27.1
SQLAlchemy~=1.4.32
bs4~=0.0.1
beautifulsoup4~=4.10.0
python-dotenv~=0.19.2
soupsieve~=2.3.1
pip~=22.0.3
Werkzeug~=2.0.3
click~=8.0.4
MarkupSafe~=2.1.0
Jinja2~=3.0.3
itsdangerous~=2.1.0
certifi~=2021.10.8
greenlet~=1.1.2
idna~=3.3
urllib3~=1.26.8
setuptools~=60.6.0
psycopg2~=2.9.3